feat(experiments): add module and add qasports2
